pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-rpa"
version = "0.1.0"
description = "AI-driven RPA system powered by Robot Framework"
requires-python = ">=3.9"
dependencies = [
    "robotframework",
    "openai>=1.0.0",
    "rich",
]

[project.scripts]
airpa = "ai_rpa.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
